% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{package}
\name{czelections}
\alias{czelections}
\alias{czelections-package}
\title{Czech Elections Data for Municipalities}
\description{
Datasets of votes in various election types in the Czech Republic, aggregated for municipalities.
}
\details{
Instead of showing election results in terms of seats won, raw data of election votes are
provided for each municipality, allowing to perform own analyses and simulations based on them.
Only a subset of election data for the independent Czech Republic is provided, beginning between
2004 and 2010 (depending on election type). Note that the data are not comprehensive since votes
given in foreign countries and votes in additional municipal elections are not included.

Two data frames correspond with each election: votes (e.g. \code{\link{ps2013}}) and turnout data
(e.g. \code{\link{ps2013_turnout}}). The names contain an abbreviation of election type as given
by CZSO: \dQuote{ep} (European Parliament), \dQuote{kv} (municipal), \dQuote{kz} (regional),
\dQuote{prez} (presidential), '\dQuote{ps} (Chamber of Deputies) or '\dQuote{se} (Senate).

There is also an overview of municipalities (\code{\link{obce}} data frame).

Elections data use data published by the Czech Statistical Office (CZSO) as opendata:
\url{http://volby.cz/opendata/opendata.htm}. They are available under the conditions listed here:
\url{https://www.czso.cz/csu/czso/conditions_for_use_and_further_dissemination}

List of municipalities uses data from Wikidata (www.wikidata.org), a free database that anyone can edit.
}

