% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{ep2004}
\alias{ep2004}
\alias{ep2009}
\alias{ep2014}
\title{European Parliament Election Votes in the Czech Republic}
\format{A data frame with 3 columns:
\describe{
  \item{party}{abbreviation of name of voted party}
  \item{votes}{number of votes given}
  \item{lau2}{municipality identifier (LAU2)}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
ep2004

ep2009

ep2014
}
\description{
Dataset containing votes for parties aggregated by municipalities.
}
\keyword{datasets}

