% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{kv2010}
\alias{kv2010}
\alias{kv2014}
\title{Municipal Election Votes in the Czech Republic}
\format{A data frame with 3 columns:
\describe{
  \item{party}{abbreviation of name of voted party; special values are \dQuote{NK} (\dQuote{nezávislý kandidát}, independent candidate)}
    and \dQuote{SNK} (\dQuote{sdružení nezávislých kandidátů}, union of independents)
  \item{votes}{number of votes given}
  \item{lau2}{municipality identifier (LAU2)}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
kv2010

kv2014
}
\description{
Dataset containing votes for parties aggregated by municipalities.
}
\keyword{datasets}

