% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{kv2010_turnout}
\alias{kv2010_turnout}
\alias{kv2014_turnout}
\title{Municipal Election Turnout in the Czech Republic}
\format{A data frame with 7 columns:
\describe{
  \item{districts}{number of polling places within the municipality}
  \item{voters}{number of all eligible voters}
  \item{envel_given}{number of envelopes given to voters}
  \item{envel_returned}{number of envelopes put into ballot box by voters}
  \item{valid_votes}{number of valid votes as counted from ballot box}
  \item{lau2}{municipality identifier (LAU2)}
  \item{seats}{number of seats to be won in the municipality}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
kv2010_turnout

kv2014_turnout
}
\description{
Election turnout data aggregated by municipalites.
}
\keyword{datasets}

