% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{kz2008}
\alias{kz2008}
\alias{kz2012}
\title{Regional Election Votes in the Czech Republic}
\format{A data frame with 3 columns:
\describe{
  \item{party}{abbreviation of name of voted party}
  \item{votes}{number of votes given}
  \item{lau2}{municipality identifier (LAU2)}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
kz2008

kz2012
}
\description{
Dataset containing votes for parties aggregated by municipalities.
}
\keyword{datasets}

