% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{obce}
\alias{obce}
\title{Municipalities of the Czech Republic}
\format{A data frame with the following columns:
\describe{
  \item{lau2}{local administrative unit of the second level (LAU2), an identifier of municipality, unique only within the Czech Republic}
  \item{name}{municipality name (with district name if needed to distinguish between the same names)}
  \item{latitude}{municipality latitude in degrees}
  \item{longitude}{municipality longitude in degrees}
  \item{lau1}{district (local administrative unit of the first level, LAU1) which municipality belongs to}
  \item{nuts3}{region (NUTS level 3) which municipality belongs to}
  \item{constituency}{identification number of constituency for Senate elections which municipality currently belongs to. (Note that
    constituency borders are adjusted according to changes in population.) NA for municipalities that includes more than one constituency
    (Praha, Brno, Ostrava and Plzeň).}
  \item{wikidata}{identifier of Wikidata item for municipality}
}}
\source{
\url{https://www.czso.cz/csu/czso/ciselnik_obci_-cisob-} (LAU2 list), \url{https://www.wikidata.org/} (other properties)
}
\usage{
obce
}
\description{
Dataset describing municipalities in the Czech Republic. The list of municipalities is valid
as of the 1st January 2015.
}
\keyword{datasets}

