% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{prez2013}
\alias{prez2013}
\title{Presidential Election Votes in the Czech Republic}
\format{A data frame with 5 columns:
\describe{
  \item{party}{abbreviation of name of party of which the candidate is member; special value \dQuote{BEZPP} (\dQuote{bez politické příslušnosti})
    means independent candidate}
  \item{candidate}{name of the candidate}
  \item{round}{election round}
  \item{votes}{number of votes given}
  \item{lau2}{municipality identifier (LAU2)}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
prez2013
}
\description{
Dataset containing votes for presidential candidates aggregated by municipalities.
}
\keyword{datasets}

