% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{ps2006}
\alias{ps2006}
\alias{ps2010}
\alias{ps2013}
\title{Chamber of Deputies Election Votes in the Czech Republic}
\format{A data frame with 3 columns:
\describe{
  \item{party}{abbreviation of name of voted party}
  \item{votes}{number of votes given}
  \item{lau2}{municipality identifier (LAU2)}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
ps2006

ps2010

ps2013
}
\description{
Dataset containing votes for parties aggregated by municipalities.
}
\keyword{datasets}

