% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{se2008}
\alias{se2008}
\alias{se2010}
\alias{se2011}
\alias{se2012}
\alias{se2014}
\title{Senate Election Votes in the Czech Republic}
\format{A data frame with 7 columns:
\describe{
  \item{party}{abbreviation of name of party that nominated the candidate}
  \item{candidate}{name of the candidate}
  \item{constituency}{identification number of constituency}
  \item{date}{election date; note that more election are held within a yeare in case of by-election}
  \item{round}{election round}
  \item{votes}{number of votes given}
  \item{lau2}{municipality identifier, usually LAU2. In case of municipality that includes more than one constituency
    (Praha, Brno, Ostrava and Plzeň), a similar code of municipal district is used instead.}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
se2008

se2010

se2011

se2012

se2014
}
\description{
Dataset containing votes for candidates aggregated by municipalities or their parts.
}
\keyword{datasets}

