% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czelections.r
\docType{data}
\name{se2008_turnout}
\alias{se2008_turnout}
\alias{se2010_turnout}
\alias{se2011_turnout}
\alias{se2012_turnout}
\alias{se2014_turnout}
\title{Senate Election Turnout in the Czech Republic}
\format{A data frame with 7 columns:
\describe{
  \item{districts}{number of polling places within the municipality}
  \item{round}{election round}
  \item{voters}{number of all eligible voters}
  \item{envel_given}{number of envelopes given to voters}
  \item{envel_returned}{number of envelopes put into ballot box by voters}
  \item{valid_votes}{number of valid votes as counted from ballot box}
  \item{lau2}{municipality identifier, usually LAU2. In case of municipality that includes more than one constituency
    (Praha, Brno, Ostrava and Plzeň), a similar code of municipal district is used instead.}
}}
\source{
\url{http://volby.cz/opendata/opendata.htm}
}
\usage{
se2008_turnout

se2010_turnout

se2011_turnout

se2012_turnout

se2014_turnout
}
\description{
Election turnout data aggregated by municipalites or their parts.
}
\keyword{datasets}

